Add support for showing/hiding elements after a point in time.
This allows us to publish a list of events, and then have them disappear from our list after the event is done.  This is still slightly buggy, in that there is a fade-out when the page loads, instead of just fade out/fade in when visibility changes after the page is loaded.